requirements.txt
ctransformers
importlib
matplotlib
networkx 
pyyaml